fix checkstyle issues
