+++ src/test/java/uk/gov/hmcts/reform/em/stitching/rest/TestUtil.java
@@ -22,14 +22,13 @@ import static org.assertj.core.api.Assertions.assertThat;
  */
 public final class TestUtil {
 
-    private TestUtil() {
-
-    }
-
     /* MediaType for JSON UTF8 */
     public static final MediaType APPLICATION_JSON_UTF8 = new MediaType(
             MediaType.APPLICATION_JSON.getType(),
             MediaType.APPLICATION_JSON.getSubtype(), StandardCharsets.UTF_8);
+
+    private TestUtil() {
+    }
 
     public static byte[] convertObjectToJsonBytes(Object object)
             throws IOException {
